Fixed Bugs #170, #138, #137, #130. Fixed #128 (minor documentation bug)
Error messages are raised for invalid date and time. Trying to schedule
a meeting in the past will raise an error message. Trying to add a
meeting on an invalid day like 31st Feb will throw an error instead of
rounding down.

Fixed #128 where there was an inconsistency in the error message for
empty time and empty description.